+++ nemo_aligner/algorithms/critic_server_trainer.py
@@ -30,12 +30,16 @@ from nemo.utils import logging
 from nemo_aligner.servers.constants import ServerSignal
 from nemo_aligner.servers.server_callables import process_inference_request
 from nemo_aligner.utils import parallel_state
-from nemo_aligner.utils.distributed import SyncTimer, broadcast_2d_tensor, rebalance_nd_tensor
+from nemo_aligner.utils.distributed import SyncTimer, broadcast_2d_tensor, run_distributed_inference
 from nemo_aligner.utils.server_utils import lock_method, pad_input
 from nemo_aligner.utils.train_utils import clip_gradients
 from nemo_aligner.utils.utils import apply_func_to_dict
 
 ENDPOINT_BIND_ADDRESS = "0.0.0.0"
+
+# once we merge the critic and actor configs
+# we should set this to num_rollout_samples in the actor
+MAX_BATCH = 9999999
 
 
 class CriticServerTrainer:
@@ -49,7 +53,7 @@ class CriticServerTrainer:
         ranks what to do
     """
 
-    def __init__(self, cfg, model, optimizer, scheduler, logger, ckpt_callback):
+    def __init__(self, cfg, model, optimizer, scheduler, logger, ckpt_callback, tokenize_func, gbs):
         self.lock = threading.Lock()
         self.logger = logger
         self.cfg = cfg
@@ -60,20 +64,30 @@ class CriticServerTrainer:
         self.gbs = cfg.gbs
         self.forward_mbs = cfg.forward_mbs
         self.step = 0
-        self.pad_sequence_length_to_multiple = cfg.pad_sequence_length_to_multiple
+
+        self.pad_sequence_length_to_multiple = cfg.get("pad_sequence_length_to_multiple", None)
+        self.tokenize_func = tokenize_func
 
         # server parameters
         self.combine_rm_and_critic_server = cfg.combine_rm_and_critic_server
+        self.max_queue_delay_microseconds = cfg.get("max_queue_delay_microseconds", 2000)
+
+        inference_micro_batch_size = cfg.inference_micro_batch_size
+        if isinstance(inference_micro_batch_size, int):  # for backwards compatability
+            inference_micro_batch_size = [inference_micro_batch_size]
+        self.preferred_batch_sizes = [
+            item * parallel_state.get_data_parallel_world_size() for item in inference_micro_batch_size
+        ]
+        self.pad_batch_to_multiple = min(self.preferred_batch_sizes)
+
         self.infer_fn = model.infer_rm_critic if self.combine_rm_and_critic_server else model.infer
         self.port = cfg.port
-        self.max_inference_batch_size = cfg.inference_micro_batch_size * parallel_state.get_data_parallel_world_size()
 
         # PyTriton args
         self.infer_inputs = (
             Tensor(name="sentences", shape=(-1,), dtype=bytes, optional=True),
             Tensor(name="tokens", shape=(-1,), dtype=np.int64, optional=True),
             Tensor(name="sequence_lengths", shape=(-1,), dtype=np.int64, optional=True),
-            Tensor(name="add_EOS", shape=(1,), dtype=np.bool_, optional=True),
         )
         self.infer_outputs = [
             Tensor(name="values", shape=(-1,), dtype=np.float32),
@@ -105,11 +119,14 @@ class CriticServerTrainer:
 
         inputs, extra, prepad_sequence_length = process_inference_request(
             inputs,
-            pad_to=self.forward_mbs * parallel_state.get_data_parallel_world_size(),
+            pad_to=self.pad_batch_to_multiple,
             pad_sequence_length_to_multiple=self.pad_sequence_length_to_multiple,
+            tokenize_func=self.tokenize_func,
         )
-        rewards, values = self.run_inference(inputs=inputs, extra=extra)
+        rewards, values = self.run_inference(inputs=inputs)
 
+        # if the inference request has extra padding that it doesn't need
+        # then we will pad it back up to the expected padding when returning the values
         if prepad_sequence_length > values.shape[1]:
             values = np.pad(
                 values, ((0, 0), (0, prepad_sequence_length - values.shape[1])), mode="constant", constant_values=0
@@ -127,7 +144,7 @@ class CriticServerTrainer:
 
     @sample
     @lock_method("self.lock")
-    def server_save(self, **inputs: np.ndarray) -> Dict[str, np.ndarray]:
+    def server_save(self, **_: np.ndarray) -> Dict[str, np.ndarray]:
         # tell other ranks to start inference
         choice = ServerSignal.SAVE.cuda()
         torch.distributed.broadcast(choice, 0)
@@ -177,18 +194,14 @@ class CriticServerTrainer:
                 http_address=ENDPOINT_BIND_ADDRESS,
                 http_port=self.port,
             )
-            # try to find a common multiple of forward mbs and dp so we don't need to pad
-            dp_size = parallel_state.get_data_parallel_world_size()
-            preferred_batch_size = [dp_size * self.forward_mbs * (i + 1) for i in range(1000)]
 
-            # 1 second latency max
             dynamic_batcher = DynamicBatcher(
-                max_queue_delay_microseconds=self.cfg.max_queue_delay_microseconds,
-                preferred_batch_size=preferred_batch_size,
+                max_queue_delay_microseconds=self.max_queue_delay_microseconds,
+                preferred_batch_size=self.preferred_batch_sizes,
             )
 
             # we cut the batch into pieces so we don't need to have a max batch size
-            infer_model_config = ModelConfig(batching=True, max_batch_size=9999999, batcher=dynamic_batcher)
+            infer_model_config = ModelConfig(batching=True, max_batch_size=MAX_BATCH, batcher=dynamic_batcher)
             # the model will split the train batch by itself
             train_model_config = ModelConfig(batching=False, max_batch_size=0, batcher=None)
             save_model_config = ModelConfig(batching=False, max_batch_size=0, batcher=None)
@@ -240,36 +253,9 @@ class CriticServerTrainer:
         """only rank 0 has valid data
         """
         self.model.prepare_for_inference()
-        tokens, lengths = None, None
-        dp_rank = parallel_state.get_data_parallel_rank()
-        dp_size = parallel_state.get_data_parallel_world_size()
-        is_rank_0 = torch.distributed.get_rank() == 0
-
-        if is_rank_0:
-            tokens = torch.as_tensor(inputs["inputs"], dtype=torch.long, device=torch.cuda.current_device())
-            lengths = torch.as_tensor(inputs["sequence_length"], dtype=torch.long, device=torch.cuda.current_device())
-
-        tokens = broadcast_2d_tensor(tokens, 0, dtype=torch.long, group=None).chunk(dp_size)[dp_rank]
-        lengths = broadcast_2d_tensor(lengths, 0, dtype=torch.long, group=None).chunk(dp_size)[dp_rank].squeeze(-1)
-
-        outputs = self.infer_fn(inputs=(tokens, lengths))
-
-        if self.combine_rm_and_critic_server:
-            rewards, values = outputs
-            rewards = (
-                rebalance_nd_tensor(rewards, group=parallel_state.get_data_parallel_group())
-                .squeeze(dim=(1,))
-                .cpu()
-                .numpy()
-            )
-
-        else:
-            values = outputs
-            rewards = None
-
-        values = rebalance_nd_tensor(values, group=parallel_state.get_data_parallel_group()).cpu().numpy()
-
+        rewards, values = run_distributed_inference(inputs, self.infer_fn, self.combine_rm_and_critic_server)
         self.model.finish_inference()
+
         torch.distributed.barrier()
         return rewards, values