+++ nemo_aligner/models/nlp/gpt/gpt_reward_model.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from abc import ABC, abstractclassmethod
 from copy import deepcopy
 from typing import Callable, List, Literal, Optional, Union
 from unittest.mock import patch
@@ -23,7 +22,6 @@ from megatron.core.models.gpt import GPTModel
 from megatron.core.tensor_parallel.layers import RowParallelLinear
 from megatron.core.transformer.spec_utils import ModuleSpec
 from megatron.core.transformer.transformer_config import TransformerConfig
-from megatron.core.utils import make_sharded_tensor_for_checkpoint, make_tp_sharded_tensor_for_checkpoint
 from torch import Tensor
 
 from nemo_aligner.utils import parallel_state