+++ nemo_aligner/utils/utils.py
@@ -354,9 +354,9 @@ def retrieve_model_state_dict_in_cpu(model, megatron_amp_O2=True):
 
 
 @torch.no_grad()
-def copy_model_states_to_cpu(model, cpu_dict=None, megatron_amp_O2=True, sync=True):
-    """NOTE: this mutates the input cpu_dict and for non tensors
-        it aliases the object
+def copy_model_states_to_cpu(model, cpu_dict=None, megatron_amp_O2=True, sync=True, alias_non_tensor=False):
+    """This function mutates the cpu_dict object to throw the model states into preallocated tensors(if they exist)
+        for non tensors it will do a deepcopy, unless alias_non_tensor is True
     """
     if cpu_dict is None:
         cpu_dict = {}
@@ -364,11 +364,12 @@ def copy_model_states_to_cpu(model, cpu_dict=None, megatron_amp_O2=True, sync=Tr
     for name, item in model.state_dict().items():
         if isinstance(item, torch.Tensor):
             if name not in cpu_dict:
-                # empty like has no pin_memory arg
                 cpu_dict[name] = torch.empty(
                     item.size(), dtype=item.dtype, layout=item.layout, device="cpu", pin_memory=True
                 )
             cpu_dict[name].copy_(item, non_blocking=False)
+        elif alias_non_tensor:
+            cpu_dict[name] = item
         else:
             cpu_dict[name] = deepcopy(item)
 
@@ -430,10 +431,10 @@ def convert_to_amp_o2_format(state_dict):
     """
     new_state_dict = {}
 
-    for key in state_dict.keys():
+    for key, item in state_dict.items():
         if "model.module." not in key:
-            new_key = key.replace("model.", "model.module.", 1)
-            new_state_dict[new_key] = state_dict[key]
+            key = key.replace("model.", "model.module.", 1)
+        new_state_dict[key] = item
 
     return new_state_dict